refactor: 추천 여행지 목록 조회 훅 리팩토링 및 서비스 함수 분리
