+++ hooks/recommendation/useFetchRecommendations.js
@@ -1,33 +1,18 @@
-import { useEffect, useState } from "react";
-import { collection, getDocs, query, orderBy } from "firebase/firestore";
-import { db } from "@/lib/firebase/client";
+"use client";
+
+import { useQuery } from "@tanstack/react-query";
+import { fetchRecommendationsFromFirestore } from "@/services/fetchRecommendations";
 
 export function useFetchRecommendations() {
-  const [recommendations, setRecommendations] = useState([]);
-  const [loading, setLoading] = useState(true);
+  const {
+    data: recommendations = [],
+    isLoading,
+    isError,
+    error,
+  } = useQuery({
+    queryKey: ["recommendations"],
+    queryFn: fetchRecommendationsFromFirestore,
+  });
 
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const q = query(
-          collection(db, "recommended_places"),
-          orderBy("createdAt", "desc")
-        );
-        const snapshot = await getDocs(q);
-        const list = snapshot.docs.map((doc) => ({
-          id: doc.id,
-          ...doc.data(),
-        }));
-        setRecommendations(list);
-      } catch (error) {
-        console.error("추천 여행지 불러오기 실패:", error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchData();
-  }, []);
-
-  return { recommendations, loading };
+  return { recommendations, isLoading, isError, error };
 }
