+++ services/fetchRecommendations.js
@@ -1,0 +1,14 @@
+import { collection, getDocs, query, orderBy } from "firebase/firestore";
+import { db } from "@/lib/firebase/client";
+
+export async function fetchRecommendationsFromFirestore() {
+  const q = query(
+    collection(db, "recommended_places"),
+    orderBy("createdAt", "desc")
+  );
+  const snapshot = await getDocs(q);
+  return snapshot.docs.map((doc) => ({
+    id: doc.id,
+    ...doc.data(),
+  }));
+}
